backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-dotenv
pydantic
python-multipart
openpyxl
pandas
Pillow
oracledb
